New Update

+++ WebApplication2/DTO/Report.cs
@@ -4,7 +4,7 @@ namespace WebApplication2.DTO;
 
 public class Report
 {
-    public float totalIncome { get; set; }
-    public float totalExpence { get; set; }
+    public double totalIncome { get; set; }
+    public double totalExpence { get; set; }
     public List<Operation> listOfCurrentOperations { get; set; }
 }
